Sqlite doesn't allow BigInteger to be used as an primary key with autoincrement

+++ gns3server/db/models/images.py
@@ -15,7 +15,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from sqlalchemy import Table, Column, String, ForeignKey, BigInteger
+from sqlalchemy import Table, Column, String, ForeignKey, BigInteger, Integer
 from sqlalchemy.orm import relationship
 
 from .base import Base, BaseTable, GUID
@@ -24,7 +24,7 @@ from .base import Base, BaseTable, GUID
 image_template_link = Table(
     "images_templates_link",
     Base.metadata,
-    Column("image_id", BigInteger, ForeignKey("images.image_id", ondelete="CASCADE")),
+    Column("image_id", Integer, ForeignKey("images.image_id", ondelete="CASCADE")),
     Column("template_id", GUID, ForeignKey("templates.template_id", ondelete="CASCADE"))
 )
 
@@ -33,7 +33,7 @@ class Image(BaseTable):
 
     __tablename__ = "images"
 
-    image_id = Column(BigInteger, primary_key=True, autoincrement=True)
+    image_id = Column(Integer, primary_key=True, autoincrement=True)
     filename = Column(String, index=True)
     path = Column(String, unique=True)
     image_type = Column(String)
